+++ server/src/Lib/Semantics/Builders/MolangSemanticTokensBuilder.ts
@@ -1,53 +1,15 @@
-import { OffsetWord } from "bc-vscode-words";
-import { SemanticTokens, SemanticTokensBuilder } from "vscode-languageserver";
-import { Position } from "vscode-languageserver-textdocument";
 import { TextDocument } from "../../Types/Document/TextDocument";
-import { SemanticModifiersEnum, SemanticTokensEnum } from "../include";
-
-//TODO refactor add base class
+import { BaseSemanticTokensBuilder } from "./BaseSemanticTokensBuilder";
 
 /**
  *
  */
-export class MolangSemanticTokensBuilder {
-  public Builder: SemanticTokensBuilder;
-  public doc: TextDocument;
-
-  constructor(doc: TextDocument) {
-    this.doc = doc;
-    this.Builder = new SemanticTokensBuilder();
-  }
-
-  Build(): SemanticTokens {
-    return this.Builder.build();
-  }
-
-  PositionAt(offset: number): Position {
-    return this.doc.positionAt(offset);
-  }
-
-  /**
-   * Adds the given text locations into the tokens builder
-   * @param startindex
-   * @param endindex
-   * @param tokenType
-   * @param tokenModifier
-   */
-  Add(startindex: number, endindex: number, tokenType: SemanticTokensEnum, tokenModifier: SemanticModifiersEnum = SemanticModifiersEnum.declaration): void {
-    const p = this.doc.positionAt(startindex);
-    const length = endindex - startindex;
-    this.Builder.push(p.line, p.character, length, tokenType, tokenModifier);
-  }
-
+export class MolangSemanticTokensBuilder extends BaseSemanticTokensBuilder {
   /**
    *
-   * @param word
-   * @param tokenType
-   * @param tokenModifier
+   * @param doc
    */
-  AddWord(word: OffsetWord, tokenType: SemanticTokensEnum, tokenModifier: SemanticModifiersEnum = SemanticModifiersEnum.declaration): void {
-    const p = this.doc.positionAt(word.offset);
-    const length = word.text.length;
-    this.Builder.push(p.line, p.character, length, tokenType, tokenModifier);
+  constructor(doc: TextDocument) {
+    super(doc);
   }
 }
