+++ server/src/Lib/Semantics/include.ts
@@ -2,5 +2,5 @@
 export * as Builders from "./Builders/include";
 export * from "./Functions";
 export * from "./Legend";
-export * from "./Molang";
+export * from "../Minecraft/Molang/Semantics";
 export * from "./OnRequest";
